Chore: move container to global components folder
