+++ src/Root.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware, combineReducers } from 'redux';
+
 import thunk from 'redux-thunk';
+import { composeWithDevTools } from 'redux-devtools-extension';
 import { createLogger } from 'redux-logger';
+
 import createHistory from 'history/createBrowserHistory';
 import {
   ConnectedRouter,
@@ -11,26 +14,20 @@ import {
 } from 'react-router-redux';
 
 import * as reducers from './store/reducers';
-import App from './containers/App';
+import App from './components/App';
 
 const history = createHistory();
+const router = routerMiddleware(history);
 
-const configureStore = () => {
-  const router = routerMiddleware(history);
-  const logger = createLogger();
+const logger = createLogger();
 
-  const store = createStore(
-    combineReducers({
-      ...reducers,
-      router: routerReducer,
-    }),
-    applyMiddleware(router, thunk, logger),
-  );
-
-  return store;
-};
-
-const store = configureStore();
+const store = createStore(
+  combineReducers({
+    ...reducers,
+    router: routerReducer,
+  }),
+  composeWithDevTools(applyMiddleware(router, thunk, logger)),
+);
 
 const Root = () => (
   <Provider store={store}>
