+++ src/components/AppHeader/AppHeader.css
@@ -45,7 +45,6 @@
 
 #dashboard-header-logo {
   width: 40px;
-  float: left;
   margin-right: 8px;
 }
