+++ src/components/Loader/Loader.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import './Loader.css';
+
+const Loader = () => <div className="loader" />;
+
+export default Loader;
