+++ src/components/Loader/index.js
@@ -1,6 +1,1 @@
-import React from 'react';
-import './Loader.css';
-
-const Loader = () => <div className="loader" />;
-
-export default Loader;
+export { default } from './Loader';
