+++ src/store/Workspaces/selectors.js
@@ -2,12 +2,12 @@ import keys from 'lodash/keys';
 
 /** Workspaces Selectors */
 
-export const isFetching = state => {
+export const isFetching = (state) => {
   const fetchingStatus = state.workspaces.isFetching;
   return fetchingStatus;
 };
 
-export const getWorkspacesById = state => {
+export const getWorkspacesById = (state) => {
   const { workspacesById } = state.workspaces;
   const workspacesByIdArray = keys(workspacesById);
