+++ src/store/reducers.js
@@ -1,3 +1,4 @@
+// App Reducers
 import workspaces from './Workspaces/reducer';
 
 export { workspaces };
